Updating metrics to pull from MOM website
Previously a local mirror on the Ubuntu Server team's server was used as
we expand this to more and more teams, it makes sense to discontinue our
local mirror and instead pull directly from the MOM site. This was also
enabled recently as my merge was accepted to produce this data and
publish it on the MOM site.