Collect per-virt/storage serials for AWS cloud images
As publication/promotion for each virt/storage combination is performed
separately (and in parallel), we should track each of those separately.
In this commit, we change the main (aws, aws-cn, aws-govcloud) cloud
serial storage to be the _oldest_ serial we find for a virt/storage
combination.  We also start storing each virt/storage combo as a cloud
named <source cloud>:<virt/storage> (e.g. aws-cn:hvm-ssd).